Fature tests api

+++ src/main/java/br/com/joao/justasimplerpgtable/model/dto/PersonagemDtoPadrao.java
@@ -10,10 +10,10 @@ public class PersonagemDtoPadrao implements Serializable{
 	public String nome;
 	public String apelido;
 	public Genero genero;
-	public Short ataque;
-	public Short defesa;
-	public Short agilidade;
-	public Short hp;
-	public Short mp;
+	public Integer ataque;
+	public Integer defesa;
+	public Integer agilidade;
+	public Integer hp;
+	public Integer mp;
 	public Clase clase;
 }
